backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
plexapi==4.15.6
python-dotenv==1.0.0
